refactor(#113926): Extracting "JsonValueOfJsonNumber" to separate file.
