+++ src/libraries/System.Text.Json/src/System.Text.Json.csproj
@@ -89,6 +89,7 @@ The System.Text.Json library is built-in as part of the shared framework in .NET
     <Compile Include="System\Text\Json\Nodes\JsonValue.CreateOverloads.cs" />
     <Compile Include="System\Text\Json\Nodes\JsonValue.cs" />
     <Compile Include="System\Text\Json\Nodes\JsonValueOfJsonBool.cs" />
+    <Compile Include="System\Text\Json\Nodes\JsonValueOfJsonNumber.cs" />
     <Compile Include="System\Text\Json\Nodes\JsonValueOfTCustomized.cs" />
     <Compile Include="System\Text\Json\Nodes\JsonValueOfT.cs" />
     <Compile Include="System\Text\Json\Nodes\JsonValueOfJsonPrimitive.cs" />
